+++ ngbatis-demo/src/main/java/ye/weicheng/ngbatis/demo/pojo/Employee.java
@@ -1,0 +1,28 @@
+package ye.weicheng.ngbatis.demo.pojo;
+
+// Copyright (c) 2022- All project authors. All rights reserved.
+//
+// This source code is licensed under Apache 2.0 License.
+
+import javax.persistence.Table;
+
+/**
+ * @author yeweicheng
+ * @since 2023-01-12 13:20
+ *   <br> Now is history!
+ */
+@Table(name = "employee")
+public class Employee extends Person {
+  public Employee() {
+  }
+
+  private String position;
+
+  public String getPosition() {
+    return position;
+  }
+
+  public void setPosition(String position) {
+    this.position = position;
+  }
+}